separate gulp tasks by file
