+++ package.json
@@ -140,6 +140,7 @@
     "publish-please": "^2.1.3",
     "recursive-copy": "^2.0.5",
     "request": "^2.58.0",
+    "require-all": "^2.0.0",
     "run-sequence": "^1.2.2",
     "saucelabs-connector": "^0.2.0",
     "serve-static": "^1.10.0",
