+++ tasks/build.js
@@ -1,0 +1,6 @@
+var gulp = require('gulp');
+
+gulp.task('build', [
+    'lint',
+    'fast-build',
+]);
