+++ tasks/clean.js
@@ -1,0 +1,8 @@
+var gulp = require('gulp');
+var del = require('del');
+
+gulp.task('clean', cleanTask);
+
+function cleanTask () {
+    return del('lib');
+}
