+++ tasks/fast-build.js
@@ -1,0 +1,9 @@
+var gulp = require('gulp');
+
+gulp.task('fast-build', [
+    'server-scripts',
+    'client-scripts',
+    'styles',
+    'images',
+    'templates',
+]);
