+++ tasks/prepare-website.js
@@ -1,0 +1,8 @@
+var gulp = require('gulp');
+
+gulp.task('prepare-website', [
+    'put-in-articles',
+    'put-in-navigation',
+    'put-in-posts',
+    'lint-docs',
+]);
