+++ tasks/test-docs-travis.js
@@ -1,0 +1,6 @@
+var gulp = require('gulp');
+
+gulp.task('test-docs-travis', [
+    'test-website-travis',
+    'lint',
+]);
