+++ tasks/test-functional-travis-legacy.js
@@ -1,0 +1,9 @@
+var gulp = require('gulp');
+var functionalTestConfig = require('../test/functional/config');
+var config = require('./config.js');
+
+gulp.task('test-functional-travis-legacy', ['build'], testFunctionalTravisLegacyTask);
+
+function testFunctionalTravisLegacyTask () {
+    return config.testFunctional('test/functional/legacy-fixtures', functionalTestConfig.testingEnvironmentNames.legacy);
+}
