+++ tasks/test-website.js
@@ -1,0 +1,8 @@
+var gulp = require('gulp');
+var config = require('./config.js');
+
+gulp.task('test-website', testWebsiteTask);
+
+function testWebsiteTask () {
+    return config.testWebsite(false);
+}
