+++ tasks/travis.js
@@ -1,0 +1,3 @@
+var gulp = require('gulp');
+
+gulp.task('travis', [process.env.GULP_TASK || '']);
